+++ src/main/html.ts
@@ -2,3 +2,4 @@ export * from './index';
 export * from './createHtmlSaxParser';
 export * from './createHtmlDomParser';
 export * from './createFromHtmlCharName';
+export * from './html-utils';
